Addressed comments by @Elchi3 regarding PR #212
