+++ locale/de.properties
@@ -106,11 +106,11 @@ http_links=HTTP Links
 # Description of test checking for HTTP links
 http_links_desc=URLs zu externen Ressourcen sollten, wenn möglich, HTTPS anstatt HTTP verwenden.
 
-# Name of test checking the explanation of data macros linking to the data source
-data_macro_explanation=Datenmakro Erklärung
+# Name of test checking the notes for data macros linking to the data source
+data_macro_note=Datenmakro Hinweis
 
-# Description of test checking the explanation of data macros linking to the data source
-data_macro_explanation_desc=Datenmakros benötigen eine Erklärung über ihnen, die einen Link zur Datenquelle enthält.
+# Description of test checking the notes for data macros linking to the data source
+data_macro_note_desc=Oberhalb von Datenmakros muss ein versteckter Hinweistext stehen, der einen Link zur Datenquelle enthält.
 
 # Name of test checking for macros with unnecessary parameters
 unnecessary_macro_params=Unnötige Makroparameter
@@ -181,11 +181,11 @@ article_length_desc=Hinweise zur Länge des Artikels
 # Error message for link using wrong locale
 link_using_wrong_locale=Link %s verwendet falsche Übersetzung. '%s' erwartet.
 
-# Error message for data macro missing an explanation above it linking to the data source
-data_macro_explanation_missing=Dem Datenmakro %s fehlt eine Erklärung, die zur Datenquelle verweist. Die Erklärung muss durch 'class="hidden"' versteckt sein.
+# Error message for data macro missing a note above it linking to the data source
+data_macro_note_missing=Dem Datenmakro %s fehlt ein Hinweistext, der zur Datenquelle verweist. Der Hinweis muss durch 'class="hidden"' versteckt sein.
 
-# Error message for data macro explanation missing a link referring to the data source
-data_macro_source_link_missing=Der Erklärung für das Datenmakro %s fehlt ein Link zur Datenquelle.
+# Error message for data macro note missing a link referring to the data source
+data_macro_source_link_missing=Dem Hinweistext für das Datenmakro %s fehlt ein Link zur Datenquelle.
 
 # Error message for macro with unnecessary parameters equalling slug
 macro_with_unnecessary_params_equalling_slug=Die Parameter des %s Makros sind unnötig, da sie aus dem Slug gelesen werden. Daher sollten sie entfernt werden.
