+++ locale/en-US.properties
@@ -106,11 +106,11 @@ http_links=HTTP links
 # Description of test checking for HTTP links
 http_links_desc=URLs to external resources should use HTTPS instead of HTTP, when possible.
 
-# Name of test checking the explanation of data macros linking to the data source
-data_macro_explanation=Data macro explanation 
+# Name of test checking the notes for data macros linking to the data source
+data_macro_note=Data macro note 
 
-# Description of test checking the explanation of data macros linking to the data source
-data_macro_explanation_desc=Data macros need an explanation above them linking to the data source.
+# Description of test checking the notes for data macros linking to the data source
+data_macro_note_desc=Data macros need a note above them linking to the data source.
 
 # Name of test checking for macros with unnecessary parameters
 unnecessary_macro_params=Unnecessary macro parameters
@@ -181,11 +181,11 @@ article_length_desc=Hints regarding the length of the article.
 # Error message for link using wrong locale
 link_using_wrong_locale=Link %s uses wrong locale. Expected '%s'.
 
-# Error message for data macro missing an explanation above it linking to the data source
-data_macro_explanation_missing=The data macro %s is missing an explanation linking to the data source. The explanation needs to be hidden using 'class="hidden"'.
+# Error message for data macro missing a note above it linking to the data source
+data_macro_note_missing=The data macro %s is missing a note linking to the data source. The note needs to be hidden using 'class="hidden"'.
 
-# Error message for data macro explanation missing a link referring to the data source
-data_macro_source_link_missing=The explanation for the data macro %s is missing a link to the data source.
+# Error message for data macro note missing a link referring to the data source
+data_macro_source_link_missing=The note for the data macro %s is missing a link to the data source.
 
 # Error message for macro with unnecessary parameters equalling slug
 macro_with_unnecessary_params_equalling_slug=The parameters of the %s macro are unnecessary, because they will be fetched from the slug, so they should be removed.
